Changed return type to Integer for movie score

+++ src/main/java/com/crejo/moviereviews/service/review/ReviewService.java
@@ -4,5 +4,5 @@ import com.crejo.moviereviews.model.Review;
 import com.crejo.moviereviews.model.User;
 
 public interface ReviewService {
-    void addReview(String movieTitle, String userName, Double movieScore);
+    void addReview(String movieTitle, String userName, Integer movieScore);
 }
